alteracoes feitas

+++ Libft/ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: mblank <mblank@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/26 12:40:28 by mblank            #+#    #+#             */
-/*   Updated: 2022/11/09 12:31:56 by mblank           ###   ########.fr       */
+/*   Updated: 2022/11/14 17:57:41 by mblank           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dst_size)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (dst_size == '\0')
